update db driver
