+++ packages/mdctl-api-driver/index.js
@@ -1,6 +1,7 @@
-require('./lib/cortex.object')
+const { CortexObject } = require('./lib/cortex.object')
 const Driver = require('./lib/driver')
 
 module.exports = {
-  Driver
+  Driver,
+  CortexObject
 }
