+++ packages/mdctl-api-driver/package.json
@@ -29,8 +29,6 @@
     "@medable/mdctl-cli": "^1.0.28",
     "@medable/mdctl-node-utils": "^1.0.0",
     "@medable/mdctl-core-utils": "^1.0.28",
-    "@medable/mdctl-credentials-provider-keychain": "^1.0.28",
-    "@medable/mdctl-credentials-provider-pouchdb": "^1.0.28",
     "async-iter-stream": "^1.2.0",
     "inflection": "^1.12.0",
     "lodash": "^4.17.14",
