+++ packages/mdctl-node-utils/package.json
@@ -26,7 +26,9 @@
   },
   "dependencies": {
     "js-yaml": "^3.13.1",
-    "create-hash": "^1.2.0"
+    "create-hash": "^1.2.0",
+    "@medable/mdctl-credentials-provider-keychain": "^1.0.28",
+    "@medable/mdctl-credentials-provider-pouchdb": "^1.0.28"
   },
   "devDependencies": {
     "chai": "^4.1.2",
